+++ Foundation/NSMethodSignature.m
@@ -98,7 +98,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
     NSUInteger promotedSize;
 
     NSGetSizeAndAlignment([self getArgumentTypeAtIndex:i],&naturalSize,&align);
-    promotedSize=((naturalSize+sizeof(int)-1)/sizeof(int))*sizeof(int);
+    promotedSize=((naturalSize+sizeof(long)-1)/sizeof(long))*sizeof(long);
 
     result+=promotedSize;
    }
